+++ prim.hh
@@ -5,17 +5,17 @@
 #include<glm/vec3.hpp>
 using namespace glm;
 
-struct aabb     { vec3 f,t; };
-struct plane    { vec3 p,n; };
-struct triangle { vec3 p,q,r; };
-struct light    { vec3 p,c; float ia,id,is; };
+struct aabb     { dvec3 f,t; };
+struct plane    { dvec3 p,n; };
+struct triangle { dvec3 p,q,r; };
+struct light    { dvec3 p,c; double ia,id,is; };
 
 struct ray {
-  vec3 o,n,n1,d;
-  float l;
-  ray(vec3,vec3);
-  void direct(float i,float j);
-  bool hit(const aabb&,float&,float&);
-  bool hit(const plane&,vec3&);
-  bool hit(const triangle&,vec3&,vec3&);
+  dvec3 o,n,n1,d;
+  double l;
+  ray(dvec3,dvec3);
+  void direct(double i,double j);
+  bool hit(const aabb&,double&,double&);
+  bool hit(const plane&,dvec3&);
+  bool hit(const triangle&,dvec3&,dvec3&);
 };
